Separate FeedbackIndicator into its own component from SigningStatus
